requirements.txt
scipy
pandas
matplotlib
tensorflow-gpu
jupyter
cartopy
opencv
pytorch
torchvision
tqdm
h5py
boto3

